Modal adaptado para mensajes de exito +restructuración de js para cruds +implementación de bootstrap en inscripción + controles sobre db y mensajes de error para php

+++ FINAL/Proyecto/app/Http/Controllers/InscripcionController.php
@@ -23,80 +23,77 @@ class InscripcionController extends Controller
     {
         // Trae socios y membresías para mostrarlos en selects
         $socios = Socio::all();
-
         $membresias = Membresia::all();
 
-        return view('inscripciones.create', compact(
-            'socios',
-            'membresias'
-        ));
+        return view('inscripciones.create', compact('socios','membresias'));
     }
 
     // Guarda una inscripción nueva
     public function store(Request $request)
     {
+        //Validación de pago inicial si esta admitido en la solicitud
+        if ($request->boolean('RegistrarPago')) {
+            $request->validate(
+                [
+                    'Monto' => 'required|integer|min:1',
+                    'FechaPago' => 'required|date',
+                    'MedioPago' => 'required'
+                ],
+                [
+                    'Monto.required' => 'Debe ingresar el monto del pago inicial.',
+                    'Monto.integer' => 'El monto debe ser un número entero.',
+                    'Monto.min' => 'El monto debe ser mayor a 0.',
+
+                    'FechaPago.required' => 'Debe ingresar la fecha del pago.',
+                    'FechaPago.date' => 'La fecha del pago no es válida.',
+
+                    'MedioPago.required' => 'Debe seleccionar un medio de pago.'
+                ]
+            );
+        }
         // Verifica si el socio ya tiene una inscripción activa
-        $existeInscripcionActiva = Inscripcion::where(
-            'SocioID',
-            $request->SocioID
-        )
-        ->where('Estado', 'Activa')
-        ->exists();
+        $existeInscripcionActiva = Inscripcion::where('SocioID',$request->SocioID)->first();
 
         // Si ya tiene una inscripción activa
         if ($existeInscripcionActiva) {
-
-            return back()->with(
-                'error',
-                'El socio ya tiene una inscripción activa.'
-            );
+            return response()->json(['success' => false, 'message' => 'El socio ya tiene una inscripción activa.'], 409);
         }
-
+        try{
         // Crear inscripción
         $inscripcion = Inscripcion::create([
-
             'SocioID' => $request->SocioID,
-
             'MembresiaID' => $request->MembresiaID,
-
             'FechaInicio' => $request->FechaInicio,
-
             'FechaVencimiento' => $request->FechaVencimiento,
-
             'Estado' => $request->Estado,
-
             'FechaCreacion' => now()
-
         ]);
 
         // Crear pago inicial si se ingresó monto
-        if ($request->Monto) {
-
+        if ($request->boolean('RegistrarPago')) {
             Pago::create([
-
                 'InscripcionID' => $inscripcion->InscripcionID,
-
                 'Monto' => $request->Monto,
-
                 'FechaPago' => str_replace(
                     'T',
                     ' ',
                     $request->FechaPago
                 ),
-
                 'MedioPago' => $request->MedioPago
-
             ]);
         }
+        }
+        catch(\Exception $e){
+            return response()->json(['success' => false, 'message' => 'Error al registrar la inscripción.', 409]);
+        }
 
-        return redirect()->route('inscripciones.index');
+        return response()->json(['success' => true]);
     }
 
     // Muestra una inscripción específica
     public function show(string $id)
     {
         $inscripcion = Inscripcion::findOrFail($id);
-
         return view('inscripciones.show', compact('inscripcion'));
     }
 
@@ -104,11 +101,8 @@ class InscripcionController extends Controller
     public function edit(string $id)
     {
         $inscripcion = Inscripcion::findOrFail($id);
-
         $socios = Socio::all();
-
         $membresias = Membresia::all();
-
         return view('inscripciones.edit', compact(
             'inscripcion',
             'socios',
@@ -119,32 +113,42 @@ class InscripcionController extends Controller
     // Actualiza una inscripción existente
     public function update(Request $request, string $id)
     {
-        $inscripcion = Inscripcion::findOrFail($id);
+        try{
+            $inscripcion = Inscripcion::findOrFail($id);
+            $inscripcion->update([
+                'SocioID' => $request->SocioID,
+                'MembresiaID' => $request->MembresiaID,
+                'FechaInicio' => $request->FechaInicio,
+                'FechaVencimiento' => $request->FechaVencimiento,
+                'Estado' => $request->Estado
+            ]);
+        }
+        catch(\Exception $e){
+            return response()->json(['success' => false, 'message' => 'Error al actualizar la inscripción.', 409]);
+        }
 
-        $inscripcion->update([
-
-            'SocioID' => $request->SocioID,
-
-            'MembresiaID' => $request->MembresiaID,
-
-            'FechaInicio' => $request->FechaInicio,
-
-            'FechaVencimiento' => $request->FechaVencimiento,
-
-            'Estado' => $request->Estado
-
-        ]);
-
-        return redirect()->route('inscripciones.index');
+        return response()->json(['success' => true]);
     }
 
     // Elimina una inscripción
     public function destroy(string $id)
     {
+        try{
         $inscripcion = Inscripcion::findOrFail($id);
 
-        $inscripcion->delete();
+        if ($inscripcion->pagos()->exists()) {
+            return response()->json([
+                'success' => false,
+                'message' => 'No se puede eliminar la inscripción porque tiene pagos asociados.'],
+                409
+            );
+        }
 
-        return redirect()->route('inscripciones.index');
+        $inscripcion->delete();
+        }
+        catch(\Exception $e){
+            return response()->json(['success' => false, 'message' => 'Error al eliminar la inscripción.', 409]);
+        }
+        return response()->json(['success' => true]);
     }
 }
